show google cast only on mobile
